requirements.txt
django==4.1.1
Pillow==9.5.0
django-debug-toolbar==4.1.0
django-simple-captcha==0.5.17